auto_runs/requirements.txt
requests==2.24.0
requests_html==0.10.0
beautifulsoup4==4.9.3
scrapingant-client==0.3.4
python-dotenv==0.19.0
pytz==2021.1